+++ changelog.md
@@ -1,8 +1,9 @@
 - v1.3.4
   - Improved bloom processing
   - Improved volumetric sunrays
-  - Added setting 'Matrix Compatibility'
+  - Added setting 'Night Atmospheric Fog Density'
   - Added setting 'Sun/Moon Weather Opacity Decrease'
+  - Added setting 'Matrix Compatibility'
   - Fixed fast leaves
   - Fixed more bugs
   - Tweaked default settings and effects